cleanup UI, show the waypoint note if present
Changed the layout a bit, to be more vertically compact:
    [ < ] system name [ > ][ X ]
          note or x/y

Show the note below the next waypoint system, if there is no note, it show the
"x/y" process text. (is a second line in the same label object)
Register the buttons with theme.
And just return the frame and let EDMC handle placement in the UI window.